+++ pure2D/src/com/funzio/pure2D/atlas/CacheAtlas.java
@@ -56,7 +56,7 @@ public class CacheAtlas extends Atlas {
         // Log.v(TAG, String.format("initBuffer(%d, %d)", mWidth, mHeight));
 
         // create a new texture
-        mFrameBuffer = new FrameBuffer(mGLState, (int) mWidth, (int) mHeight, false);
+        mFrameBuffer = new FrameBuffer(mGLState, mWidth, mHeight, false);
         mTexture = (BufferTexture) mFrameBuffer.getTexture();
     }
