+++ pure2D/src/com/funzio/pure2D/atlas/ImageSequenceBufferAtlas.java
@@ -128,7 +128,7 @@ public class ImageSequenceBufferAtlas extends Atlas {
         mHeight = Pure2DUtils.getNextPO2(height);
 
         // create a new texture
-        mFrameBuffer = new FrameBuffer(mGLState, (int) mWidth, (int) mHeight, false);
+        mFrameBuffer = new FrameBuffer(mGLState, mWidth, mHeight, false);
         mTexture = (BufferTexture) mFrameBuffer.getTexture();
     }
